refactor(tui): retire the test-only resume snapshot renderer
render_resume_picker, _preview_lines, and _wrap had no production callers:
the picker renders through the shared browser, and the transcript owns the
preview viewport. They moved into the test package as
resume_picker_snapshot.py (a deterministic fixture importing the production
clip/fragment helpers), so the thirteen snapshot assertions keep running
against the same model state without shipping dead rendering code.

The superseded model scroll state went with it: ResumePickerModel.scroll_preview
and every preview_offset write (the live picker scrolls the transcript
directly). The model scroll test now exercises the real surface — scrolling
moves the transcript top row and row changes are visible.
